refactor: move server to main layer
